Add files via upload
This Unity script, ARController, is designed for use with AR Foundation to manage and respond to changes in tracked images within an augmented reality (AR) scene. It primarily focuses on handling events when AR images are detected, updated, or removed from the scene. Upon detecting a new tracked image, the script dynamically assigns a specified prefab as a child to the detected image's transform, allowing developers to easily associate 3D objects or UI elements with specific real-world images recognized by the AR session. This functionality is crucial for creating interactive AR experiences that overlay digital content on real-world images, such as educational tools, gaming, and marketing applications. The script ensures a seamless integration of digital content with physical spaces, enhancing user engagement and immersion in AR environments.